refactor: isolate profile onboarding administration
Changed:
Moved config loading and saving, revision fallback, preview session creation, preview streaming, and preview restart state into an admin controller hook.

Benefit:
The admin page now focuses on rendering while preserving unsaved-config previews and existing compatibility behavior.
